# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      tags: [News]
      operationId: listNewsItems
      summary: List News
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of News.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemList'
    post:
      tags: [News]
      operationId: createNewsItem
      summary: Create a NewsItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsItemInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItem'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /news/{id}:
    patch:
      tags: [News]
      operationId: updateNewsItem
      summary: Update a NewsItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsItemPatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItem'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [News]
      operationId: deleteNewsItem
      summary: Delete a NewsItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /gallery-images:
    get:
      tags: [Gallery]
      operationId: listGalleryImages
      summary: List Gallery
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Gallery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryImageList'
    post:
      tags: [Gallery]
      operationId: createGalleryImage
      summary: Create a GalleryImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryImageInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryImage'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /gallery-images/{id}:
    patch:
      tags: [Gallery]
      operationId: updateGalleryImage
      summary: Update a GalleryImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GalleryImagePatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryImage'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Gallery]
      operationId: deleteGalleryImage
      summary: Delete a GalleryImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /testimonials:
    get:
      tags: [Testimonials]
      operationId: listTestimonials
      summary: List Testimonials
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Testimonials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestimonialList'
    post:
      tags: [Testimonials]
      operationId: createTestimonial
      summary: Create a Testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /testimonials/{id}:
    patch:
      tags: [Testimonials]
      operationId: updateTestimonial
      summary: Update a Testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestimonialPatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testimonial'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Testimonials]
      operationId: deleteTestimonial
      summary: Delete a Testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /videos:
    get:
      tags: [Videos]
      operationId: listVideos
      summary: List Videos
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Videos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoList'
    post:
      tags: [Videos]
      operationId: createVideo
      summary: Create a Video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /videos/{id}:
    patch:
      tags: [Videos]
      operationId: updateVideo
      summary: Update a Video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoPatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Videos]
      operationId: deleteVideo
      summary: Delete a Video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /programmes:
    get:
      tags: [Programmes]
      operationId: listProgrammes
      summary: List Programmes
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Programmes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeList'
    post:
      tags: [Programmes]
      operationId: createProgramme
      summary: Create a Programme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgrammeInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Programme'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /programmes/{id}:
    patch:
      tags: [Programmes]
      operationId: updateProgramme
      summary: Update a Programme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgrammePatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Programme'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Programmes]
      operationId: deleteProgramme
      summary: Delete a Programme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /stats:
    get:
      tags: [Stats]
      operationId: listStats
      summary: List Stats
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Stats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatList'
    post:
      tags: [Stats]
      operationId: createStat
      summary: Create a Stat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stat'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /stats/{id}:
    patch:
      tags: [Stats]
      operationId: updateStat
      summary: Update a Stat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatPatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stat'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Stats]
      operationId: deleteStat
      summary: Delete a Stat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /values:
    get:
      tags: [Values]
      operationId: listSchoolValues
      summary: List Values
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Values.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolValueList'
    post:
      tags: [Values]
      operationId: createSchoolValue
      summary: Create a SchoolValue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolValueInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolValue'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /values/{id}:
    patch:
      tags: [Values]
      operationId: updateSchoolValue
      summary: Update a SchoolValue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolValuePatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolValue'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Values]
      operationId: deleteSchoolValue
      summary: Delete a SchoolValue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admission-steps:
    get:
      tags: [Admissions]
      operationId: listAdmissionSteps
      summary: List Admissions
      parameters:
        - name: includeHidden
          in: query
          required: false
          schema:
            type: boolean
          description: Admin-only. When true and authenticated, include hidden items.
      responses:
        '200':
          description: List of Admissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionStepList'
    post:
      tags: [Admissions]
      operationId: createAdmissionStep
      summary: Create a AdmissionStep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdmissionStepInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionStep'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /admission-steps/{id}:
    patch:
      tags: [Admissions]
      operationId: updateAdmissionStep
      summary: Update a AdmissionStep
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdmissionStepPatch'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionStep'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Admissions]
      operationId: deleteAdmissionStep
      summary: Delete a AdmissionStep
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    NewsItem:
      type: object
      required: [id, message, isVisible, sortOrder, createdAt]
      properties:
        id:
          type: integer
        message:
          type: string
        isVisible:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    NewsItemList:
      type: array
      items:
        $ref: '#/components/schemas/NewsItem'
    NewsItemInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
    NewsItemPatch:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        isVisible:
          type: boolean
        sortOrder:
          type: integer
    GalleryImage:
      type: object
      required: [id, caption, category, objectPath, wide, isVisible, sortOrder, createdAt]
      properties:
        id:
          type: integer
        caption:
          type: string
        category:
          type: string
        objectPath:
          type: string
        wide:
          type: boolean
        isVisible:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    GalleryImageList:
      type: array
      items:
        $ref: '#/components/schemas/GalleryImage'
    GalleryImageInput:
      type: object
      required: [caption, objectPath]
      properties:
        caption:
          type: string
          minLength: 1
        category:
          type: string
        objectPath:
          type: string
          minLength: 1
        wide:
          type: boolean
    GalleryImagePatch:
      type: object
      properties:
        caption:
          type: string
          minLength: 1
        category:
          type: string
        objectPath:
          type: string
          minLength: 1
        wide:
          type: boolean
        isVisible:
          type: boolean
        sortOrder:
          type: integer
...
    TestimonialList:
      type: array
      items:
        $ref: '#/components/schemas/Testimonial'
    TestimonialInput:
      type: object
      required: [quote, name]
      properties:
        quote:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        role:
          type: string
        initials:
          type: string
...
    VideoList:
      type: array
      items:
        $ref: '#/components/schemas/Video'
...